update some error
