+++ app/src/main/java/com/space/quizapp/presentation/auth/AuthenticationFragment.kt
@@ -1,15 +1,11 @@
 package com.space.quizapp.presentation.auth
 
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
 import androidx.navigation.fragment.findNavController
 import com.space.quizapp.common.extensions.collectFlow
 import com.space.quizapp.databinding.FragmentAuthenticationBinding
-import com.space.quizapp.presentation.base.BaseFragment
+import com.space.quizapp.presentation.base.fragment.BaseFragment
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class AuthenticationFragment :
@@ -24,7 +20,7 @@ class AuthenticationFragment :
         }
     }
 
-    override fun observes() {
+    override fun setObserves() {
         collectFlow(viewModel.successNavigation) {
             findNavController().navigate(AuthenticationFragmentDirections.actionGlobalHomeFragment())
         }
